update_db changes + new category column
